requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=eigen
absl-py=0.11.0=pyhd3eb1b0_1
adal=1.2.5=pypi_0
aiohttp=3.6.3=py37he774522_0
alembic=1.4.1=pypi_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py37_0
attrs=20.3.0=pyhd3eb1b0_0
azure-common=1.1.26=pypi_0
azure-core=1.10.0=pypi_0
azure-graphrbac=0.61.1=pypi_0
azure-mgmt-authorization=0.61.0=pypi_0
azure-mgmt-containerregistry=2.8.0=pypi_0
azure-mgmt-keyvault=2.2.0=pypi_0
azure-mgmt-resource=12.0.0=pypi_0
azure-mgmt-storage=11.2.0=pypi_0
azure-storage-blob=12.7.0=pypi_0
azureml-core=1.20.0=pypi_0
backports-tempfile=1.0=pypi_0
backports-weakref=1.0.post1=pypi_0
bcrypt=3.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py37_0
boto3=1.16.54=pypi_0
botocore=1.19.54=pypi_0
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2020.12.8=haa95532_0
cachetools=4.2.0=pyhd3eb1b0_0
certifi=2020.12.5=py37haa95532_0
cffi=1.14.4=py37hcd4344a_0
chardet=4.0.0=pypi_0
click=7.1.2=py_0
cloudpickle=1.6.0=pypi_0
contextlib2=0.6.0.post1=pypi_0
cryptography=3.3.1=py37hcd4344a_0
cycler=0.10.0=py37_0
databricks-cli=0.14.1=pypi_0
docker=4.4.1=pypi_0
entrypoints=0.3=pypi_0
flask=1.1.2=pypi_0
freetype=2.10.4=hd328e21_0
gast=0.3.3=py_0
gitdb=4.0.5=pypi_0
gitpython=3.1.12=pypi_0
google-api-core=1.24.1=pypi_0
google-auth=1.24.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-cloud-core=1.5.0=pypi_0
google-cloud-storage=1.35.0=pypi_0
google-crc32c=1.1.0=pypi_0
google-pasta=0.2.0=py_0
google-resumable-media=1.2.0=pypi_0
googleapis-common-protos=1.52.0=pypi_0
grpcio=1.31.0=py37he7da953_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
imdbclassifier=0.6.6=pypi_0
importlib-metadata=3.4.0=pypi_0
intel-openmp=2020.2=254
isodate=0.6.0=pypi_0
itsdangerous=1.1.0=pypi_0
jeepney=0.6.0=pypi_0
jinja2=2.11.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jsonpickle=1.4.2=pypi_0
keras=2.4.3=0
keras-applications=1.0.8=py_1
keras-base=2.4.3=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.0=py37hd77b12b_0
kubernetes=12.0.1=pypi_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0.1=h200bbdf_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
mako=1.1.3=pypi_0
markdown=3.3.3=py37haa95532_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py37hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mleap=0.16.1=pypi_0
mlflow=1.13.1=pypi_0
msrest=0.6.19=pypi_0
msrestazure=0.6.4=pypi_0
multidict=4.7.6=py37he774522_1
ndg-httpsclient=0.5.1=pypi_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
openssl=1.1.1i=h2bbff1b_0
opt_einsum=3.1.0=py_0
pandas=1.2.0=pypi_0
paramiko=2.7.2=pypi_0
pathspec=0.8.1=pypi_0
pillow=8.1.0=py37h4fa10fc_0
pip=20.3.3=py37haa95532_0
prometheus-client=0.9.0=pypi_0
prometheus-flask-exporter=0.18.1=pypi_0
protobuf=3.14.0=pypi_0
pyarrow=2.0.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pyjwt=1.7.1=py37_0
pynacl=1.4.0=pypi_0
pyopenssl=19.1.0=pypi_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pysftp=0.2.9=pypi_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python-editor=1.0.4=pypi_0
pytz=2020.5=pypi_0
pywin32=227=pypi_0
pyyaml=5.3.1=pypi_0
qt=5.9.7=vc14h73c81de_0
querystring-parser=1.2.4=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7=pyhd3eb1b0_1
ruamel-yaml=0.16.12=pypi_0
ruamel-yaml-clib=0.2.2=pypi_0
s3transfer=0.3.4=pypi_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
secretstorage=3.3.0=pypi_0
setuptools=51.1.2=py37haa95532_4
sip=4.19.8=py37h6538335_0
six=1.15.0=py37haa95532_0
smmap=3.0.4=pypi_0
sqlalchemy=1.3.22=pypi_0
sqlite=3.33.0=h2a8f88b_0
sqlparse=0.4.1=pypi_0
tabulate=0.8.7=pypi_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py37h04bc1aa_0
tensorflow-base=2.3.0=eigen_py37h17acbac_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
termcolor=1.1.0=py37_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
tornado=6.1=py37h2bbff1b_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
waitress=1.4.4=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.6.3=py37h2bbff1b_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
